requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
python-dotenv==1.0.1
pytest==7.4.2
